Created database with relations[Finishes #159884254]

+++ app/api/v1/endpoints.py
@@ -1,25 +1,4 @@
-'''api endpoints:
-    /api/v1/questions
-        methods=['GET']
-        returns json data of all the questions
-        status_code 200
-    /api/v1/get_questions/<int:question_id>
-        methods=['GET']
-        returns json data on the question specified
-        status_code 200
-    /api/v1/post_question
-        methods=['POST']
-        returns json data on the added question
-        status_code 201
-    /api/v1/questions/<int:question_id>/answers
-        methods=['POST']
-        returns json data on the question answered
-        status_code 201
-
-errors:
-    404 resource not found
-    201 created
-    200 OK
+'''api endpoints
 '''
 
 from datetime import datetime as dt
@@ -61,6 +40,7 @@ def get_question(question_id):
     sql = 'SELECT * FROM answers WHERE question_id=%s'
     cursor.execute(sql, [question_id])
     answers = cursor.fetchall()
+    cursor.close()
     return jsonify({'question': question}, {'answers':answers}), 200
 
 
@@ -69,10 +49,10 @@ def get_question(question_id):
 def add_question():
     ''' add a question
     '''
-    if request.json and request.json['owner'] and request.json['content']:
+    if request.json and request.json['content']:
         cursor = CONNECTION.cursor()
         sql = 'INSERT INTO questions(content, question_owner) VALUES (%s, %s);'
-        cursor.execute(sql, (request.json['content'], request.json['owner']))
+        cursor.execute(sql, (request.json['content'], int(current_identity)))
         CONNECTION.commit()
         cursor.close()
         return jsonify({'201': 'Question added'}), 201
@@ -91,9 +71,9 @@ def answer_question(question_id):
     question = cursor.fetchall()
     if not question:
         return abort(404), 404
-    if request.json and request.json['answer_content'] and request.json['answer_owner']:
+    if request.json and request.json['answer_content']:
         sql = 'INSERT INTO answers(answer_owner, content, question_id) VALUES (%s, %s, %s);'
-        cursor.execute(sql,(request.json['answer_owner'], request.json['answer_content'], question_id))
+        cursor.execute(sql,(int(current_identity), request.json['answer_content'], question_id))
         CONNECTION.commit()
         cursor.close()
         return jsonify({"201": "question answered"}), 201
@@ -110,10 +90,58 @@ def remove_question(question_id):
     question = cursor.fetchall()
     if not question:
         return abort(404), 404
-    sql = 'DELETE FROM questions WHERE id=%s;'
-    cursor.execute(sql, ([question_id]))
-    sql = 'DELETE FROM answers WHERE question_id=%s;'
-    cursor.execute(sql, ([question_id]))
-    cursor.close()
-    CONNECTION.commit()
-    return jsonify({'Question deleted':'200'}), 200
+    if int(current_identity) is question[0][2]:
+        sql = 'DELETE FROM questions WHERE id=%s;'
+        cursor.execute(sql, ([question_id]))
+        sql = 'DELETE FROM answers WHERE question_id=%s;'
+        cursor.execute(sql, ([question_id]))
+        cursor.close()
+        CONNECTION.commit()
+        return jsonify({'Question deleted':'200'}), 200
+    return jsonify({"401":"Unauthorized: Only question owner can remove question"})
+
+@BP.route('/upvote/<int:answer_id>', methods=['POST'])
+@jwt_required()
+def upvote_answer(answer_id):
+    '''Upvote answers
+    '''
+    cursor = CONNECTION.cursor()
+    sql = 'SELECT * FROM answers WHERE id=%s;'
+    cursor.execute(sql, ([answer_id]))
+    answer = cursor.fetchall()
+    if answer:
+        try:
+            if str(current_identity) in answer[0][7]:
+                return jsonify({"400":"You can only vote once"})
+        except TypeError:
+            pass
+        answer[0] = list(answer[0])
+        answer[0][3] +=1
+        sql = 'SELECT array_append(voters[], %s);'
+        cursor.execute(sql, ([str(current_identity)]))
+        answer[0] = tuple(answer[0])
+        CONNECTION.commit()
+        cursor.close()
+    return abort(404)
+
+@BP.route('/downvote/<int:answer_id>', methods=['POST'])
+@jwt_required()
+def downvote_answer(answer_id):
+    '''Downvote answers
+    '''
+    cursor = CONNECTION.cursor()
+    sql = 'SELECT * FROM answers WHERE id=%s;'
+    cursor.execute(sql, ([answer_id]))
+    answer = cursor.fetchall()
+    if answer:
+        try:
+            if str(current_identity) in answer[0][7]:
+                return jsonify({"400":"You can only vote once"})
+        except TypeError:
+            pass
+        answer[0][3] -= 1;
+        sql = 'SELECT array_append(%s, %s);'
+        cursor.execute(sql, (answer[0][7], str(current_identity)))
+        CONNECTION.commit()
+        cursor.close()
+    return abort(404)
